Broke up implicits module into submodules, ignored lift and zipped
